Added headers
Moved ob_start into it's own function so we can hook it into wp_loaded which ensures all headers are present before sending the buffer and header over to the blob cahce handler. Testing confirmed that some headers are not sent when the ob_start() call is directly inside of init().